requirements.txt
Django
djangorestframework
ultralytics
Pillow